+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Robbery optimisation".
 - Tests for "The voucher".
 - Tests for "Connect the hyper-dots".
+- Tests for "Rock-paper-scissors war".
 
 ## [1.12.0] - 2022-09-01
 ### Added